Move to_16bit and clip signal to clarity/utils
