+++ clarity/utils/signal_processing.py
@@ -37,3 +37,36 @@ def denormalize_signals(sources: np.ndarray, ref: np.ndarray) -> np.ndarray:
     Returns:
         np.ndarray: Signal rescaled back to its original."""
     return sources * ref.std() + ref.mean()
+
+
+def clip_signal(signal: np.ndarray, soft_clip: bool = False) -> tuple[np.ndarray, int]:
+    """Clip the signal.
+
+    Args:
+        signal (np.ndarray): Signal to be clipped and saved.
+        soft_clip (bool): Whether to use soft clipping.
+
+    Returns:
+        signal (np.ndarray): Clipped signal.
+        n_clipped (int): Number of samples clipped.
+    """
+
+    if soft_clip:
+        signal = np.tanh(signal)
+    n_clipped = np.sum(np.abs(signal) > 1.0)
+    signal = np.clip(signal, -1.0, 1.0)
+    return signal, int(n_clipped)
+
+
+def to_16bit(signal: np.ndarray) -> np.ndarray:
+    """Convert the signal to 16 bit.
+
+    Args:
+        signal (np.ndarray): Signal to be converted.
+
+    Returns:
+        signal (np.ndarray): Converted signal.
+    """
+    signal = signal * 32768.0
+    signal = np.clip(signal, -32768.0, 32767.0)
+    return signal.astype(np.dtype("int16"))
