+++ tests/utils/test_signal_processing.py
@@ -1,11 +1,14 @@
 """Test for utils.signal_processing module"""
+# pylint: disable=import-error
 import numpy as np
 import pytest
 
 from clarity.utils.signal_processing import (
+    clip_signal,
     compute_rms,
     denormalize_signals,
     normalize_signal,
+    to_16bit,
 )
 
 
@@ -135,3 +138,38 @@ def test_compute_rms():
     assert rms == pytest.approx(
         57.803515840, rel=pytest.rel_tolerance, abs=pytest.abs_tolerance
     )
+
+
+@pytest.mark.parametrize(
+    "signal,soft_clip,expected_output",
+    [
+        (
+            np.array([0.5, 2.0, -1.5, 0.8]),
+            True,
+            (np.array([0.46211716, 0.96402758, -0.90514825, 0.66403677]), 0),
+        ),
+        (np.array([0.5, 2.0, -1.5, 0.8]), False, (np.array([0.5, 1.0, -1.0, 0.8]), 2)),
+    ],
+)
+def test_clip_signal(signal, soft_clip, expected_output):
+    """Test the clip_signal function"""
+    # Test with soft clip
+    output = clip_signal(signal, soft_clip=soft_clip)
+    assert np.allclose(output[0], expected_output[0])
+    assert output[1] == expected_output[1]
+
+
+@pytest.mark.parametrize(
+    "signal,expected_output",
+    [
+        (np.array([0.5, 0.8, 0.2, 1.0]), np.array([16384, 26214, 6553, 32767])),
+        (np.array([-0.5, -0.8, -0.2, -1.0]), np.array([-16384, -26214, -6553, -32768])),
+        (np.array([0.5, -0.8, 0.2, -1.0]), np.array([16384, -26214, 6553, -32768])),
+    ],
+)
+def test_to_16bit(signal, expected_output):
+    """Test the to_16bit function"""
+    # Test with positive signal
+    output = to_16bit(signal)
+    print(output)
+    assert np.allclose(output, expected_output)
